requirements.txt
Flask==2.3.3 
Flask-SQLAlchemy==3.1.1
Flask-Migrate 
#flask_validator 
psycopg2-binary
python-dotenv  
pymupdf
python-docx
textract
langchain
langchain_groq